[detailed-metrics] extract existing DB counter helpers
